Moved files to be consistent with this version of CMSSW
